% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{get_config}
\alias{get_config}
\title{Get config file options}
\usage{
get_config(config_file = "slushy_config.yml")
}
\arguments{
\item{config_file}{yml file containing the configs. Defaults to `slushy_config.yml` in current working directory. If set to `NULL`, then config file from slushy package will be used.}
}
\value{
list of config options and values
}
\description{
Get config file options
}
