% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{new_config}
\alias{new_config}
\title{Create a new slushy_config.yml from template in project directory}
\usage{
new_config(config = NULL, path = getwd())
}
\arguments{
\item{config}{Name of config to default to}

\item{path}{Path to store slushy_config.yml. Defaults to current working directory}
}
\description{
Create a new slushy_config.yml from template in project directory
}
