% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slushy_add.R
\name{slushy_add}
\alias{slushy_add}
\title{Add packages to an existing slushy environment.}
\usage{
slushy_add(pkgs, project = NULL, config = get_config(), restart = TRUE)
}
\arguments{
\item{pkgs}{Character vector of packages to install.}

\item{project}{The project directory. If NULL, defaults to the nearest parent
directory that contains an `.Rproj` file relative to the current working directory.}

\item{config}{List of configuration options. If not provided, defaults to reading from default settings in config file included in slushy package.}

\item{restart}{Restart session after package is added? Defaults to TRUE.}
}
\value{
This function is called for its side effects.
}
\description{
Install and document additional packages in a slushy environment.
}
\examples{
\dontrun{
  slushy_add("newpkg")
}
}
