% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slushy_add.R
\name{slushy_drop}
\alias{slushy_drop}
\title{Drop packages from an existing slushy environment.}
\usage{
slushy_drop(pkgs, project = NULL, restart = TRUE)
}
\arguments{
\item{pkgs}{Character vector of packages to remove}

\item{project}{The project directory. If NULL, defaults to the nearest parent
directory that contains an `.Rproj` file relative to the current working directory.}

\item{restart}{Restart session after package is added? Defaults to TRUE.}
}
\value{
This function is called for its side effects.
}
\description{
Drop packages from an existing slushy environment.
}
\examples{
\dontrun{
  slushy_drop("oldpkg")
}
}
