% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slushy_init.R
\name{slushy_reinit}
\alias{slushy_reinit}
\title{Re-Initialize a slushy environment}
\usage{
slushy_reinit(project = NULL, restart = TRUE, config = get_config())
}
\arguments{
\item{project}{The project directory. If NULL, defaults to the nearest parent
directory that contains an `.Rproj` file relative to the current working directory.}

\item{restart}{Restart session after slushy is initialized? Defaults to TRUE.}

\item{config}{List of configuration options. If not provided, defaults to reading from default settings in config file included in slushy package.}
}
\value{
This function is called for its side effects.
}
\description{
Following an update of the {slushy} package, re-initialize the slushy
environment using the current list of packages, snapshot date, and
environment settings.
}
