% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slushy_remove.R
\name{slushy_remove}
\alias{slushy_remove}
\title{Remove slushy infrastructure from a project.}
\usage{
slushy_remove(project = NULL, restart = TRUE, keep = "slushy_config.yml")
}
\arguments{
\item{project}{The project directory. If NULL, defaults to the nearest parent
directory that contains an `.Rproj` file relative to the current working directory.}

\item{restart}{Restart session after slushy is removed? Defaults to TRUE.}
}
\value{
This function is called for its side effects.
}
\description{
Deletes the files and folders needed for the slushy environment before
restarting.
}
\examples{
\dontrun{
 slushy_remove()
}
}
