% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slushy_sync.R
\name{slushy_sync}
\alias{slushy_sync}
\title{Sync local environment to match the project's lockfile.}
\usage{
slushy_sync(project = NULL)
}
\arguments{
\item{project}{The project directory. If NULL, defaults to the nearest parent
  directory that contains an `.Rproj` file relative to the current working directory.

Install/update packages as needed to match the lockfile.}
}
\value{
This function is called for its side effects.
}
\description{
Sync local environment to match the project's lockfile.
}
\examples{
\dontrun{
  slushy_sync()
}
}
