% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slushy_trim.R
\name{slushy_trim}
\alias{slushy_trim}
\title{Trim installed packages}
\usage{
slushy_trim(
  project = NULL,
  config = get_config(config_file = "slushy_config.yml")
)
}
\arguments{
\item{project}{The project directory. If NULL, defaults to the nearest parent
directory that contains an `.Rproj` file relative to the current working directory.}

\item{config}{List of configuration options. If not provided, defaults to reading from 
default settings in config file included in slushy package.}
}
\description{
This function will trim down the installed packages to only those that are 
used in the code files of the project. Unused packages will be removed from 
the DESCRIPTION file, uninstalled from the project library, and removed from 
the `renv.lock` file.
}
