% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slushy_update.R
\name{slushy_update_preview}
\alias{slushy_update_preview}
\title{Preview the update to a slushy environment}
\usage{
slushy_update_preview(date = NULL, project = NULL, config = get_config())
}
\arguments{
\item{date}{CRAN snapshot date to use for packages. If NULL, today's date will be used.}

\item{project}{The project directory. If NULL, defaults to the nearest parent
directory that contains an `.Rproj` file relative to the current working directory.}

\item{config}{List of configuration options. If not provided, defaults to reading from default settings in config file included in slushy package.}
}
\value{
Console message describing updates to be made.
}
\description{
Preview the update to a slushy environment
}
\examples{
\dontrun{

 slushy_update_preview()
}

}
